+++ typescript/21-leetcode-trees /max-depth.ts
@@ -1,13 +1,4 @@
-export class _Node {
-  val: number;
-  children: _Node[];
-
-  constructor(val?: number, children?: _Node[]) {
-    this.val = val === undefined ? 0 : val;
-    this.children = children === undefined ? [] : children;
-  }
-}
-
+import { _Node } from "./trees.js";
 
 export function maxDepth(root: _Node | null): number {
   if (!root) return 0;
