+++ typescript/21-leetcode-trees /same-tree.ts
@@ -1,15 +1,4 @@
-export class TreeNode {
-  val: number
-  left: TreeNode | null
-  right: TreeNode | null
-
-  constructor(val?: number, left?: TreeNode | null, right?: TreeNode | null) {
-    this.val = (val === undefined ? 0 : val)
-    this.left = (left === undefined ? null : left)
-    this.right = (right === undefined ? null : right)
-  }
-}
-
+import { TreeNode} from "./trees.js"
 
 export function isSameTree(p: TreeNode | null, q: TreeNode | null): boolean {
   if (p === null && q === null) return true;
