+++ typescript/21-leetcode-trees /trees.ts
@@ -9,3 +9,13 @@ export class TreeNode {
     this.right = (right === undefined ? null : right)
   }
 }
+
+export class _Node {
+  val: number;
+  children: _Node[];
+
+  constructor(val?: number, children?: _Node[]) {
+    this.val = val === undefined ? 0 : val;
+    this.children = children === undefined ? [] : children;
+  }
+}
